Lint and pylint config
