+++ setup.cfg
@@ -18,3 +18,8 @@ column_limit = 120
 
 [pycodestyle]
 max-line-length = 120
+
+[pylint]
+disable = too-many-arguments,
+          too-few-public-methods,
+          broad-except
